+++ mml/examples/plain.rs
@@ -1,9 +1,11 @@
+#[cfg(feature = "async-std")]
+use async_std::main;
 use mml::MmlCompilerBuilder;
+#[cfg(feature = "tokio")]
+use tokio::main;
 
-#[tokio::main]
+#[test_log::test(main)]
 async fn main() {
-    env_logger::builder().is_test(true).init();
-
     let mml = include_str!("./plain.eml");
     let mml_compiler = MmlCompilerBuilder::new().build(mml).unwrap();
     let mime = mml_compiler.compile().await.unwrap().into_string().unwrap();